requirements.txt
pypdf
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
requests
python-dotenv
